moved some components to container folder
