+++ react/components/PublishUrlMiddle.jsx
